+++ CHANGELOG.md
@@ -18,6 +18,9 @@ This release significantly improves the runtime performance of Metro-generated g
 - **Enhancement**: Don't generate provider fields for graph self instances unless necessary.
 - **Enhancement**: Improve accuracy of diagnostic location when reporting graph validation issues from binding callable declarations.
 - **Enhancement**: Validate bindings from interop with Dagger `@BindsOptionalOf` lazily.
+- **Enhancement**: [graph analysis] Title graph edges with semantically meaningful descriptions of the edge type. They should read a bit more like plain English.
+- **Enhancement**: [graph analysis] Better handle visualizing bound instance inputs.
+- **Enhancement**: [graph analysis] Better visualize member injection from graph injectors.
 - **Fix**: Ensure IC lookups are recorded for all merged contributions.
 - **Fix**: Fix `newInstance()` args not stripping `Lazy` in top-level function inject classes.
 - **Fix**: Allow `Any` to be a `binding<...>()` type if no explicit supertypes are declared.